Refactor routes into submodules
