+++ app/__init__.py
@@ -47,9 +47,13 @@ def create_app():
         return get_user(user_id)
 
     # Register blueprints
-    from .routes import bp as main_bp
+    from .routes.guest_routes import guest_bp
+    from .routes.animal_routes import animal_bp
+    from .routes.payment_routes import payment_bp
 
-    app.register_blueprint(main_bp)
+    app.register_blueprint(guest_bp)
+    app.register_blueprint(animal_bp)
+    app.register_blueprint(payment_bp)
     from .auth import auth_bp
 
     app.register_blueprint(auth_bp)
