+++ app/templates/admin/edit_settings.html
@@ -86,7 +86,7 @@
 
 		</fieldset >
 		<button type="submit" class="btn btn-success">Änderungen speichern</button>
-		<a href="{{ url_for('main.index') }}" class="btn btn-secondary ms-2">Abbrechen</a>
+                <a href="{{ url_for('guest.index') }}" class="btn btn-secondary ms-2">Abbrechen</a>
 
 	</form>
 </div>
