+++ app/templates/list_guests.html
@@ -40,14 +40,14 @@
 					{% endif %}
 				</td>
 				<td>
-					<a href="{{ url_for('main.view_guest', guest_id=gast.id) }}" class="btn btn-info btn-sm">
+					<a href="{{ url_for('guest.view_guest', guest_id=gast.id) }}" class="btn btn-info btn-sm">
 						<i class="fas fa-eye" title="Gast ansehen"></i>
 					</a>
-					<a href="{{ url_for('main.edit_guest', guest_id=gast.id) }}" class="btn btn-warning btn-sm">
+					<a href="{{ url_for('guest.edit_guest', guest_id=gast.id) }}" class="btn btn-warning btn-sm">
 						<i class="fas fa-pen" title="Gast bearbeiten"></i>
 					</a>
 					{% if current_user.role in ['admin', 'editor'] %}
-					<a href="{{ url_for('main.register_animal', guest_id=gast.id) }}" class="btn btn-success btn-sm">
+					<a href="{{ url_for('animal.register_animal', guest_id=gast.id) }}" class="btn btn-success btn-sm">
 						<i class="fas fa-plus" title="Tier hinzufügen"></i>
 					</a>
 					{% endif %}
@@ -74,11 +74,11 @@
 							<p class="text-danger small">Löschen kann nicht rückgängig gemacht werden!</p>
 						</div>
 						<div class="modal-footer">
-							<form action="{{ url_for('main.deactivate_guest', guest_id=gast.id) }}" method="post"
+							<form action="{{ url_for('guest.deactivate_guest', guest_id=gast.id) }}" method="post"
 								  style="display:inline;">
 								<button type="submit" class="btn btn-warning">Nur deaktivieren</button>
 							</form>
-							<form action="{{ url_for('main.delete_guest', guest_id=gast.id) }}" method="post"
+							<form action="{{ url_for('guest.delete_guest', guest_id=gast.id) }}" method="post"
 								  style="display:inline;">
 								<button type="submit" class="btn btn-danger">Vollständig löschen</button>
 							</form>
@@ -131,14 +131,14 @@
 						{% endif %}
 					</td>
 					<td>
-					<a href="{{ url_for('main.view_guest', guest_id=gast.id) }}" class="btn btn-info btn-sm">
+					<a href="{{ url_for('guest.view_guest', guest_id=gast.id) }}" class="btn btn-info btn-sm">
 						<i class="fas fa-eye" title="Gast ansehen"></i>
 					</a>
-					<a href="{{ url_for('main.edit_guest', guest_id=gast.id) }}" class="btn btn-warning btn-sm">
+					<a href="{{ url_for('guest.edit_guest', guest_id=gast.id) }}" class="btn btn-warning btn-sm">
 						<i class="fas fa-pen" title="Gast bearbeiten"></i>
 					</a>
 					{% if current_user.role in ['admin', 'editor'] %}
-					<a href="{{ url_for('main.register_animal', guest_id=gast.id) }}" class="btn btn-success btn-sm">
+					<a href="{{ url_for('animal.register_animal', guest_id=gast.id) }}" class="btn btn-success btn-sm">
 						<i class="fas fa-plus" title="Tier hinzufügen"></i>
 					</a>
 					{% endif %}
@@ -165,11 +165,11 @@
 								<p class="text-danger small">Löschen kann nicht rückgängig gemacht werden!</p>
 							</div>
 							<div class="modal-footer">
-								<form action="{{ url_for('main.activate_guest', guest_id=gast.id) }}" method="post"
+								<form action="{{ url_for('guest.activate_guest', guest_id=gast.id) }}" method="post"
 									  style="display:inline;">
 									<button type="submit" class="btn btn-warning">Gast aktivieren</button>
 								</form>
-								<form action="{{ url_for('main.delete_guest', guest_id=gast.id) }}" method="post"
+								<form action="{{ url_for('guest.delete_guest', guest_id=gast.id) }}" method="post"
 									  style="display:inline;">
 									<button type="submit" class="btn btn-danger">Vollständig löschen</button>
 								</form>
